+++ tests/portscan/0002.sh
@@ -1,5 +1,5 @@
 logdir="$(mktemp -dt portscan-test.XXXXXXX)"
-${PORTSCAN} -o unknown-variables -p 0002 -l "${logdir}/log"
+${PORTSCAN} --unknown-variables -p 0002 -l "${logdir}/log"
 [ "$(readlink "${logdir}/log/portscan-latest.log")" != "$(readlink "${logdir}/log/portscan-previous.log")" ]
 cat <<EOF | diff -u - "${logdir}/log/portscan-latest.log"
 V       archivers/arj                            IGNORE_PATCHES