+++ tests/portscan/0005.sh
@@ -1,5 +1,5 @@
 logdir="$(mktemp -dt portscan-test.XXXXXXX)"
-${PORTSCAN} -o categories -p 0005 -l "${logdir}/log"
+${PORTSCAN} --categories -p 0005 -l "${logdir}/log"
 [ "$(readlink "${logdir}/log/portscan-latest.log")" != "$(readlink "${logdir}/log/portscan-previous.log")" ]
 cat <<EOF | diff -u - "${logdir}/log/portscan-latest.log"
 C       archivers                                unsorted category or other formatting issues